fix according to review

+++ internal/migrations/sql/3_create_table_withdrawals.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE withdrawals
     id UUID DEFAULT gen_random_uuid() PRIMARY KEY,
     create_at TIMESTAMPTZ NOT NULL,
     order_num BIGINT NOT NULL UNIQUE,
-    sum FLOAT,
+    sum DECIMAL,
     user_id UUID NOT NULL,
     CONSTRAINT fk_customer
         FOREIGN KEY(user_id)
